Moving code around for cleanup
